Support ERL_ROOTDIR passed as an environment variable
Update the Erlang/OTP releases for Android to include the change discussed
at https://github.com/erlang/otp/pull/2863 for the erl and start scripts.

This allows to set dynamically the absolute path of the Erlang runtime,
as it is configured in a different location for multiple users on Android
(/data/data/... for the main user but /data/user/ID/... for other users).